+++ tests/test_common.h
@@ -11,13 +11,13 @@
 #include "include/assert.h"
 
 void fill_buf(double *buf, size_t buf_size) {
-    for (int i = 0; i < buf_size; ++i) {
+    for (size_t i = 0; i < buf_size; ++i) {
         buf[i] = (double) i;
     }
 }
 
 void assert_buf(const double *exp, const double *real, size_t size, double tol) {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         double a = exp[i];
         double b = real[i];
         LWTEST_ASSERT_ALMOST_EQUAL_DOUBLE(a, b, tol);
