+++ tests/test_repart.c
@@ -5,8 +5,8 @@
 
 #include "test_common.h"
 
-void test_reshape(caterva_ctx_t *ctx, uint64_t ndim, uint64_t *shape_, uint64_t *pshape_,
-                  uint64_t *pshape_dest_) {
+void test_reshape(caterva_ctx_t *ctx, int8_t ndim, int64_t *shape_, int64_t *pshape_,
+                  int64_t *pshape_dest_) {
 
     caterva_dims_t shape = caterva_new_dims(shape_, ndim);
     caterva_dims_t pshape = caterva_new_dims(pshape_, ndim);
@@ -14,7 +14,7 @@ void test_reshape(caterva_ctx_t *ctx, uint64_t ndim, uint64_t *shape_, uint64_t 
 
     caterva_array_t *src = caterva_empty_array(ctx, NULL, pshape);
 
-    uint64_t buf_size = 1;
+    size_t buf_size = 1;
     for (int i = 0; i < CATERVA_MAXDIM; ++i) {
         buf_size *= (shape.dims[i]);
     }
@@ -28,11 +28,11 @@ void test_reshape(caterva_ctx_t *ctx, uint64_t ndim, uint64_t *shape_, uint64_t 
 
     caterva_repart(dest, src);
 
-    double *buf_dest = (double *) malloc(dest->size * src->sc->typesize);
+    double *buf_dest = (double *) malloc((size_t)dest->size * src->sc->typesize);
 
     caterva_to_buffer(dest, buf_dest);
 
-    assert_buf(buf_src, buf_dest, dest->size, 1e-14);
+    assert_buf(buf_src, buf_dest, (size_t)dest->size, 1e-14);
     free(buf_src);
     free(buf_dest);
     caterva_free_array(src);
@@ -53,28 +53,28 @@ LWTEST_TEARDOWN(reshape) {
 }
 
 LWTEST_FIXTURE(reshape, ndim2) {
-    const uint64_t ndim = 2;
-    uint64_t shape_[] = {100, 100};
-    uint64_t pshape_[] = {10, 10};
-    uint64_t pshape_dest_[] = {20, 20};
+    const int8_t ndim = 2;
+    int64_t shape_[] = {100, 100};
+    int64_t pshape_[] = {10, 10};
+    int64_t pshape_dest_[] = {20, 20};
 
     test_reshape(data->ctx, ndim, shape_, pshape_, pshape_dest_);
 }
 
 LWTEST_FIXTURE(reshape, ndim2_n) {
-    const uint64_t ndim = 2;
-    uint64_t shape_[] = {8563, 8234};
-    uint64_t pshape_[] = {356, 353};
-    uint64_t pshape_dest_[] = {1033, 1033};
+    const int8_t ndim = 2;
+    int64_t shape_[] = {8563, 8234};
+    int64_t pshape_[] = {356, 353};
+    int64_t pshape_dest_[] = {1033, 1033};
 
     test_reshape(data->ctx, ndim, shape_, pshape_, pshape_dest_);
 }
 
 LWTEST_FIXTURE(reshape, ndim6) {
-    const uint64_t ndim = 6;
-    uint64_t shape_[] = {20, 15, 18, 14, 19, 20};
-    uint64_t pshape_[] = {2, 5, 4, 9, 4, 12};
-    uint64_t pshape_dest_[] = {3, 6, 8, 13, 15, 11};
+    const int8_t ndim = 6;
+    int64_t shape_[] = {20, 15, 18, 14, 19, 20};
+    int64_t pshape_[] = {2, 5, 4, 9, 4, 12};
+    int64_t pshape_dest_[] = {3, 6, 8, 13, 15, 11};
 
     test_reshape(data->ctx, ndim, shape_, pshape_, pshape_dest_);
 }
